+++ .storybook/main.js
@@ -1,0 +1,14 @@
+/** @type { import('@storybook/html-vite').StorybookConfig } */
+const config = {
+  stories: ['../stories/**/*.stories.@(js|jsx|ts|tsx)'],
+  addons: ['@storybook/addon-essentials'],
+  framework: {
+    name: '@storybook/html-vite',
+    options: {},
+  },
+  docs: {
+    autodocs: 'tag',
+  },
+};
+
+export default config;